requirements.txt
app==0.0.1
appdirs==1.4.3
arrow==0.15.5
asgiref==3.2.3
attrs==18.2.0
autopep8==1.4.4
beautifulsoup4==4.8.2
binaryornot==0.4.4
boto==2.48.0
bson==0.5.8
cached-property==1.5.1
camelcase==0.2
certifi==2019.3.9
chardet==3.0.7
Click==7.0
colorama==0.4.3
comtypes==1.1.7
configparser==4.0.2
cookiecutter==1.7.0
crayons==0.3.0
cycler==0.10.0
dataclasses==0.6
defusedxml==0.6.0
diskcache==2.9.0
distlib==0.3.0
Django==3.0.6
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-mongodb-engine-py3==0.6.0.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.2
dnspython==1.16.0
falcon==2.0.0
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gunicorn==20.0.4
html-testRunner==1.2.1
idna==2.8
imdbpie==5.6.4
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jinja2-time==0.2.0
joblib==0.14.1
kiwisolver==1.2.0
linecache2==1.0.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mongoengine==0.9.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.0.1
pandas==1.0.3
pep8==1.7.1
Pillow==7.0.0
pipenv==2018.11.26
poyo==0.5.0
protobuf==3.6.1
psutil==5.7.0
psycopg2==2.8.5
pycodestyle==2.5.0
pymongo==3.10.1
pyparsing==2.4.7
pypiwin32==223
pyspeech==0.0.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2018.5
pywin32==227
pywinauto==0.6.8
PyYAML==5.3
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
scikit-learn==0.22.1
scipy==1.4.1
selenium==3.141.0
serial==0.0.97
six==1.14.0
soupsieve==2.0
speech==0.5.2
SpeechRecognition==3.8.1
sqlparse==0.2.4
suds-py3==1.3.4.0
tinycom==1.1
traceback2==1.4.0
trans==2.1.0
unittest2==1.1.0
urllib3==1.24.2
vboxapi==1.0
virtualenv==20.0.3
virtualenv-clone==0.5.3
waitress==1.4.3
webdriver-manager==2.3.0
Werkzeug==1.0.0
whichcraft==0.6.1
xlrd==1.2.0
zeep==3.4.0
